Add init and check datapackage tasks + refactoring
