+++ data_quality/generators/base.py
@@ -16,12 +16,14 @@ class BaseGenerator(object):
         if not self.base_url:
             raise TypeError('Cannot generate the database without the "url" parameter.')
 
-    def generate_sources(self, sources_filepath, file_types=['csv','excel']):
+    def generate_sources(self, sources_filepath, file_types=['csv', 'excel']):
         """Generate sources file for CSV database"""
 
-        raise NotImplementedError('You must overwrite this method with your generator\'s specific logic.')
+        raise NotImplementedError('You must overwrite this method with your '
+                                  'generator\'s specific logic.')
 
     def generate_publishers(self, publishers_filepath):
         """Generate publishers file for CSV database"""
 
-        raise NotImplementedError('You must overwrite this method with your generator\'s specific logic.')
+        raise NotImplementedError('You must overwrite this method with your '
+                                  'generator\'s specific logic.')
