+++ data_quality/utilities.py
@@ -4,22 +4,21 @@ from __future__ import division
 from __future__ import print_function
 from __future__ import unicode_literals
 
-import os
 import io
+import os
 import json
 import shutil
-import pkg_resources
 import collections
 import datapackage
-from . import compat
+import pkg_resources
 
 def set_up_cache_dir(cache_dir_path):
     """Reset /cache_dir before a new batch."""
 
     if os.path.lexists(cache_dir_path):
         for root, dirs, files in os.walk(cache_dir_path):
-            for file in files:
-                os.unlink(os.path.join(root, file))
+            for contained_file in files:
+                os.unlink(os.path.join(root, contained_file))
 
             for directory in dirs:
                 shutil.rmtree(os.path.join(root, directory))
@@ -44,19 +43,8 @@ def resolve_dir_name(config_filepath, dir_path):
     if not os.path.isabs(dir_path):
         config_path = os.path.abspath(os.path.dirname(config_filepath))
         return os.path.join(config_path, dir_path)
-
-def get_resource_by_name(name, datapackage):
-    """Retrieve resource from datapackage by name"""
-
-    resources = [resource for resource in datapackage.resources
-                 if resource.metadata['name'] == name]
-
-    if len(resources) == 1:
-        return resources[0]
-    elif len(resources) == 0:
-        return None
     else:
-        raise ValueError('There is more than one resource with this name')
+        return dir_path
 
 def load_json_config(config_filepath):
     """Loads the json config into a dictionary, overwriting the defaults"""
@@ -66,8 +54,8 @@ def load_json_config(config_filepath):
 
     if not config_filepath:
         return default_config
-    with io.open(config_filepath, mode='rt', encoding='utf-8') as file:
-        user_config = json.loads(file.read())
+    with io.open(config_filepath, mode='rt', encoding='utf-8') as config_file:
+        user_config = json.loads(config_file.read())
         config = deep_update_dict(default_config, user_config)
         config['data_dir'] = resolve_dir_name(config_filepath, config['data_dir'])
         config['cache_dir'] = resolve_dir_name(config_filepath, config['cache_dir'])
@@ -81,29 +69,20 @@ def get_default_datapackage():
     datapkg = datapackage.DataPackage(json.loads(default_datapkg.decode('utf-8')))
     return datapkg
 
-def load_json_datapackage(config):
-    """Generate a datapackage or return the existing one along with it's path"""
+def get_datapackage_resource(resource_path, datapkg):
+    """Return the resource correspondent to `resource_path` from datapackage or raise"""
 
-    datapkg_filepath = config.get('datapackage_file', '')
-
-    if not datapkg_filepath or not os.path.isabs(datapkg_filepath):
-        data_dir_path = os.path.normpath(config['data_dir'])
-        datapkg_dir_path = os.path.dirname(data_dir_path)
-        datapkg_filepath = os.path.join(datapkg_dir_path, 'datapackage.json')
-
-    datapkg_filepath = os.path.abspath(datapkg_filepath)
-
-    if not os.path.exists(datapkg_filepath):
-        with io.open(datapkg_filepath, mode='w+', encoding='utf-8') as new_datapkg:
-            default_datapkg = get_default_datapackage()
-            for resource in default_datapkg.resources:
-                resource.metadata['path'] = os.path.join(config['data_dir'],
-                                                         resource.metadata['path'])
-            json_datapkg = json.dumps(default_datapkg.to_dict(), indent=4)
-            new_datapkg.write(compat.str(json_datapkg))
-            return (default_datapkg, datapkg_filepath)
-
-    return  (datapackage.DataPackage(datapkg_filepath), datapkg_filepath)
+    matching_resources = [res for res in datapkg.resources
+                          if res.local_data_path == resource_path]
+    if len(matching_resources) > 1:
+        raise ValueError(('The resource with path "{0}" appears multiple times '
+                          'in your datapackage.').format(resource_path))
+    elif not matching_resources:
+        raise ValueError(('The resource with path "{0}" can\'t be found in '
+                          'your datapackage. Please include it or '
+                          'use the "dq init" command.').format(resource_path))
+    else:
+        return matching_resources[0]
 
 def deep_update_dict(source_dict, new_dict):
     """Update a nested dictionary (modified in place) with another dictionary.
